moved openocd config, hooks in other file
